perf(crypto/math): optimize mul_by_6 and mul_by_7 for Goldilocks extensions
- mul_by_6: Changed from 8a - 2a (3 doubles + 1 sub) to 4a + 2a
  (2 doubles + 1 add), saving one double operation
- mul_by_7: Changed from 8a - a (3 doubles + 1 sub) to a + 2a + 4a
  (2 doubles + 2 adds), saving one double operation

Benchmarks show 57% improvement in Fp3 multiplication throughput.

Also moved extension field tests from inline module to dedicated
test file (goldilocks_asm_tests.rs) for better organization.
